REFACTOR / 폴더 정리
레이드 관련 interface raid/interface에 정리
사용하지 않던 common 디렉토리 삭제

Resolve: #76
