+++ src/raid/interface/enterBossOption.interface.ts
