+++ src/raid/interface/raidStatus.interface.ts
