+++ src/raid/interface/rankingInfo.interface.ts
